Update README.md
This project aims to provide a comprehensive example of building and deploying a CNN-based image classifier. With detailed explanations and a practical GUI interface, it serves as a valuable resource for both beginners and experienced developers interested in deep learning and computer vision. Whether for educational purposes, research, or real-world applications, this project offers a user-friendly and customizable solution for image classification tasks.